+++ test/managedTests.hs
@@ -72,9 +72,9 @@ makeWithWrapperWithMessage ref inMsg outMsg v handler = do
 
 managedCauldron :: IORef [Text] -> Cauldron Managed
 managedCauldron ref =
-  fromRecipeList
-    [ recipe @(Logger IO) $ eff $ wire $ managed (makeLogger ref),
-      recipe @(Weird IO)
+  mconcat
+    [ singleton @(Logger IO) $ eff $ wire $ managed (makeLogger ref),
+      singleton @(Weird IO)
         Recipe
           { bean = eff do
               wire \logger self -> managed (makeSelfInvokingWeird ref logger self),
@@ -83,7 +83,7 @@ managedCauldron ref =
                 [ val $ wire makeWeirdDecorator
                 ]
           },
-      recipe @(Logger IO, Weird IO) $ val_ do wire (,)
+      singleton @(Logger IO, Weird IO) $ val_ do wire (,)
     ]
 
 tests :: TestTree